Add proof of concept rewrite
Using the built-in Exec type as a starting point, and renaming
parameters

(The renamed parameters are:
  :command   => :query
  :retries   => :max_retries
  :try_sleep => :polling_frequency)

Then the Exec :returns property becomes the Wait_for :exit_code, which
then becomes a property instead of a command.

Polling_frequency is changed to accept Floats (and looking at the code,
it would be trivial to fix the Exec type so that it too accepted
Floats!)

I get the refreshonly functionality working.

To test, create these two files:

  # /tmp/exit_code.sh
  #!/bin/bash
  sleep 5
  exit 42

and

  # /tmp/test.pp
  notify { 'foo': }
  ~>
  wait_for { '/tmp/exit_code.sh':
    exit_code   => 42,
    refreshonly => true,
  }

Then:

  $ bundle exec puppet apply --modulepath=spec/fixtures/modules \
/tmp/test.pp