Change regex to type to a string
Prior to this commit, the regex parameter was munged into a Regexp. This
caused pcore serialization of the Regexp in a report, when the regex was
really just a string used for comparison within a regex. This commit
changes the regex to be a string and uses the string as a regex when
testing for a match.